requirements.txt
alembic>=1.13.3
psycopg>=3.2.2
sqlalchemy[asyncio]>=2.0.35
sqlalchemy-utils>=0.41.2
python-dotenv>=1.0.1
pytest>=8.3.3
pytest-asyncio>=0.24.0
